Create ps2
This is the VHDL to connect the keyboard and the DE2_115 board. First, register the generated data from the keyboard. When detect the keyboard's clock falling edge, then get the data sample.  It need to capture the key code value on falling edge. Then come out the corresponding keyboard signal to the board.