Create vgacode
This is the VHDL to drive the VGA. The VGA controller uses a 640x480 resolution with a 25.175Hz refresh rate, which is smaller than the maximum resolution of the monitor (1280x1024) used. All these pixels are provided directly by the FPGA. This defines the starting coordinates of a block of dynamic in VGA, and the pixels of each block. The controller contains two counters. One counter controls the horizontal timing signal, and the other controls the vertical timing signal. In this design,  the image is 3*3  blocks, and each has 30 pixels. With the dynamic blocks, first, judge the coordinates and update the coordinates then the graphics box to yellow.