Game Control VHDL Code
The game control code includes the time control part, the scorekeeper part, and the dynamic yellow blocks part. In the scorekeeper part, dividing the two digit number into two parts. In the dynamic blocks part, control the blocks with the random numbers. Using the loop and clocks to control the circuits.