VGA display code
The VGA controller uses a 640x480 resolution with a 25.175Hz refresh rate. The controller contains two counters. One counter controls the horizontal timing signal, and the other controls the vertical timing signal. In this design, the image is 3x3 blocks, and each has 30 pixels. We referred to the VHDL code from the eewiki and revised the resolution to 640x480 in order to suit our lab monitor. Adding the images and the elements that we need in our 3x3 matrix.